requirements.txt
pyyaml
torch~=2.0.1
torchvision~=0.15.2 
hydra-core~=1.3.0
lightning~=2.0.8
torchmetrics~=1.1.1
tensorboard~=2.14.0
kaggle~=1.5.16
dvc~=3.19.0
wandb~=0.15.10
fsspec~=2023.9.0
scikit-learn~=1.3.0
--extra-index-url https://download.pytorch.org/whl/cu118